Add alias to merge the current branch into another branch, omitting any conflicts or actually merging files
That is, the HEAD tree (~ working copy contents) will be declared as the tree
of the branch given as a parameter. Files will not be merged, so there will be
no conflicts. Be aware that any changes on the branch you merge into will be
lost unless they are also reachably by HEAD before the merge.

Thanks to @danielmarbach for the first draft of this alias!