zsh: update-dotfiles now supports a custom shell wrapper
This is useful for systems where zsh is not the default login shell,
e.g. a Synology NAS with zsh installed from synocommunity.

In this case, define your ~/.dotfiles-hosts as follows:
synology.example.com /usr/local/bin/zsh -c

For other hosts, where zsh is the default login shell, only define the
hostname:
host.with.zsh.as.default.example.com

This update drops support for custom commands to be run after a the
dotfiles update.